Remove usage of square brackets in test IDs
Some test IDs are dynamically generated at runtime based on a number of properties of the objects under test.
In these cases, the test ID might contain a square-bracket delimited array, such as:

propValues=[a b]

These brackets interfere with regex matching when using -skip and -run command line parameters, since brackets are reserved
for character classes in regex.

This commit provides a new formatSliceOrArray helper which formats without the brackets, and includes commas for
readability.