Add Roku support (#130)
Limitations with the Roku platform require the SDKs to connect to
alternative streaming endpoints. These endpoints communicate using
custom encryption, which we do not wish to replicate here.

Instead, the Roku SDK will connect to the same alternative endpoint, but
the information will be transmitted in plain text. The SDK can be
configured to operate in plain text mode, which should only be used for
the contract tests.