feat: Allow configuring the query timeout. (#180)
In flutter we don't really have a way to pre-build and then run the
contract tests. So we try to run them and sometimes it takes more than
10 seconds. This just lets you configure that 10 seconds and still
defaults to 10 seconds.

The limitation on the flutter side is because we need to run headless
and we are using the unit test runner to accomplish this. Not ideal, but
flutter isn't really a headless thing overall.