feat: support testing TLS options with two new capabilities (#208)
This PR adds two new capabilities. The first is `tls:skip-verify-peer`.
If provided, then the SDK under test will receive:
* `skipVerifyPeer bool`: whether or not the peer should be verified

This allows for tests of streaming/polling/events under the https
protocol using a self-signed certificate.

The second is `tls:verify-peer`. If provided, the harness will make the
SDK under test connect using https and present a self-signed
certificate. This should be rejected by the SDK.