feat: add custom CA capability (#215)
This adds a new capability `tls:custom-ca`. SDKs advertising this
capability will be passed a path to a root CA bundle.

This bundle should be used when creating a TLS connection. The harness
uses this capability to test that the SDK can trust a self-signed
certificate.

This has been tested with C++ (server/client) and Erlang Server SDKs.