chore: Introduce opt-in flag to enable persistence support (#250)
The SDK test harness has historically been self-sufficient. With the
introduction of persistence-based unit tests, that is no longer the
case. Externally started services are required for each persistence
technology supported.

Since we want to retain the ease of use associated with a our current
`go run .` usage, we are going to disable persistence support by
default. Users will have to provide a `-enable-persistence-tests` if
they want the test harness to attempt those tests.

SDKs under test are still required to provide the appropriate
capabilities. In the event those capabilities are provided, but the test
harness flag is not provided, persistence tests will be skipped.