chore: Introduce DataSystem configuration (#259)
This commit introduces the concept of the data system configuration. In
addition, all existing polling and streaming configurations have been
updated to be expressed in terms of primary synchronizers.

As it currently stands, most of the tests are passing. There are a
handful that fail which I believe are expected. Even so, this is going
into a feature branch, so I'm not too concerned about the failures just
yet.