chore: Remove top level streaming, polling, and service endpoint configurations (#260)
With the new FDv2 setup, we no longer have top level streaming and
polling configurations. Everything is configured through the new data
system configuration.

Because of this change, supporting service endpoints as a concept no
longer really applies. It is totally reasonable to think you might want
to initialize a poller from a different URL than the synchronizer. So
having a single top level configuration for those endpoints doesn't
align with the future structure.

So we are removing all of that.