chore: Elevate payload filter to top level data system property
In FDv1, the payload filter was defined directly on the streaming or
polling configuration. While  redundant then, it becomes even more so
with the FDv2 change.

To ensure we are thinking about the payload filter in the same way the
SDK handles applying it, we are moving the filter up to act as a data
system wide option.