fix: Only use h.RequireNever, not require.Never. (#305)
The `require.Never` method uses goroutines internally which results in
race conditions with assertions that are using http requests. In order
to prevent this we have `h.RequireNever` which is a replacement that
does condition checks synchronously.

The tests that used `require.Never` would periodically fail. When
running these tests locally the rate of failure had become quite high.
It is unclear what changed to make the failures more pronounced, but the
tests are stabilized with this change.